cell.c created
