+++ src/build32.sh
@@ -8,5 +8,5 @@ echo ./gen mem.s
 ./gen mem.s
 fi
 
-echo gcc -g3 lisp.c -o lisp
-gcc -m32 -g3 lisp.c -o lisp
+echo gcc -g3 lisp.c cell.c -o lisp
+gcc -m32 -g3 lisp.c cell.c -o lisp
